Reorganize marginalize codebase
